fix(studio): count composer characters like the server
The client counter measured raw markdown length while the server
counts rendered text (remark-word-limit semantics), so bodies with
links showed red counters that still posted, and vice versa.

Extract the shared rules (countMemoChars, limits, MemoSummary,
toCreatedAt) into memo-format.ts — fs- and DOM-free like
memo-filter.ts — and import them from both sides. This also removes
the drifting duplicate MemoSummary/limit definitions in client.ts,
and gives toCreatedAt its first tests.
